ecommerce_v2/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
python-dotenv==1.0.1
Flask-Login==0.6.3
bcrypt==4.2.0
stripe==10.5.0
requests==2.32.3
